port: native macOS/Apple Silicon support for libcbm + Astrolabe
Astrolabe now builds and runs natively on aarch64-apple-darwin. Indexing works
end to end: real graphs are extracted, persisted, and read back from SQLite.

Where macOS offers a better primitive than the Win32 idiom being mirrored, this
uses the better one rather than transliterating:

- source_snapshot capture uses APFS fclonefileat() copy-on-write instead of a
  byte-for-byte copy. Cloning from the already-open, already-identity-verified
  descriptor also closes a path-reopen race the Windows branch avoids with a
  retained handle. Measured: capturing 250 files wrote 0 bytes to disk.
- The SQLite writer's durable sync uses F_FULLFSYNC, the only Darwin call that
  waits for the drive to commit to stable media; plain fsync() does not.
- Store verification proves "the bytes verified are the bytes used" by exact
  before/after family identity (dev, ino, size, mtime, ctime) instead of the
  Windows freeze-and-snapshot-copy, and so verifies the source in place.

Cross-process coordination uses flock(), whose kernel-on-death release
reproduces WAIT_ABANDONED crash-safety that POSIX named semaphores lack (macOS
has neither robust mutexes nor sem_timedwait). The honest asymmetry is recorded
in the code: a crashed owner's lock is indistinguishable from a free one, so
ABANDONED/recovered states are never asserted on POSIX.

Blockers resolved:
- pipeline_internal.h: unguarded windows.h cascaded to 24 translation units
- pipeline.c: Win32 process probe -> proc_pid_rusage/getrusage
- cli.c: _NSGetEnviron used without <crt_externs.h>
- store.c: Win32 ERROR_* in host-neutral paths; POSIX in-place verification
- mcp.c: named-mutex transition lock, index admission, and store quiescence
  (quiescence now probes SQLite's own locking protocol, the authority on
  whether another process holds the store)
- source_snapshot.c: deliberate #error replaced with a Darwin implementation
- extract_imports.c / sqlite_writer.c: MSVC CRT spellings behind cbm_* wrappers
- Makefile.cbm: -fvisibility=hidden hid the entire cbm_* API on Mach-O and
  collapsed the objcopy export list to 12 symbols (see #900); linker response
  file is now Windows-only, since Apple ld64 has no @file syntax
- cbm-sys: bindings are committed per target; MSVC and clang disagree on the
  underlying type of an unsigned C enum
- astrolabe-fleet: Job Objects -> process groups via pre_exec(setpgid)
- calyx-aster / calyx-forge: Darwin process accounting; statvfs width

Refs #895, #896, #897, #898, #899, #900

+++ calyx/crates/calyx-aster/src/pressure.rs
@@ -276,9 +276,11 @@ impl Drop for TempFile {
 fn os_disk_sample(path: &Path) -> Result<DiskSample> {
     let stat = nix::sys::statvfs::statvfs(path)
         .map_err(|error| io_error(format!("statvfs {}: {error}", path.display())))?;
+    // statvfs block counts are 32-bit on Darwin and 64-bit on Linux; widen so
+    // one DiskSample representation serves both.
     Ok(DiskSample {
-        blocks: stat.blocks(),
-        blocks_available: stat.blocks_available(),
+        blocks: u64::from(stat.blocks()),
+        blocks_available: u64::from(stat.blocks_available()),
     })
 }
